Added new application structure and i18n
