+++ i18n/locales/ru.json
@@ -1,0 +1,28 @@
+{
+  "weather": {
+    "feelsLike": "Ощущается как",
+    "windSpeed": "Скорость ветра",
+    "rainChance": "Вероятность дождя",
+    "humidity": "Влажность",
+    "rainy": "Дождливо",
+    "reload": "Обновить",
+    "temperatureRange": "Диапазон температур"
+  },
+  "date": {
+    "weekdayShort": "чт.",
+    "monthShort": "апр."
+  },
+  "city": "Минск",
+  "clock": {
+    "weatherDescriptions": [
+      "Ясно", "Облачно", "Небольшой дождь", "Ливень",
+      "Гроза", "Снег", "Туман", "Пасмурно",
+      "Переменная облачность", "Морось", "Град", "Метель"
+    ],
+    "directions": ["С", "В", "Ю", "З"],
+    "windSpeedUnit": "м/с",
+    "dateFormat": "D.M",
+    "today": "Сегодня",
+    "tomorrow": "Завтра"
+  }
+}
