+++ package.json
@@ -10,16 +10,18 @@
   },
   "dependencies": {
     "@react-native-community/blur": "^4.4.1",
-    "expo": "~52.0.41",
+    "expo": "^52.0.44",
     "expo-blur": "^14.0.3",
     "expo-font": "~13.0.4",
     "expo-linear-gradient": "~14.0.2",
     "expo-splash-screen": "~0.29.22",
     "expo-status-bar": "~2.0.1",
+    "i18next": "^24.2.3",
     "lottie-ios": "^4.5.1",
     "lottie-react-native": "^7.2.2",
     "nativewind": "^4.1.23",
     "react": "18.3.1",
+    "react-i18next": "^15.4.1",
     "react-native": "^0.76.8",
     "react-native-safe-area-context": "4.12.0",
     "react-native-svg": "^15.8.0",
